(SERVER-325) Refactor `borrow-from-pool*`
In order to add support for the `max-requests-per-instance` setting,
we are going to need to make the `borrow-from-pool*` function
recursive in the case where a borrowed instance has exceeded the
maximum number of requests.  This commit refactors the function
to make it more amenable to the recursion, which will be added
in a subsequent commit.