(SERVER-325) Separate JRuby schemas to namespace
The implementations of the borrow-from-pool logic that we need
to build in order to support flushing individual JRuby instances
after they've handled a maximum number of requests was going to
create some circular dependencies between namespaces based on
the current organization.

This commit moves all of the schemas, types, and related predicates
into their own new leaf namespace as a first step towards
re-organizing things in such a way as to prevent the circular
dependencies.