(MAINT) be more tolerant of ssl exceptions in tests
Because the HTTP client library we are using is asynchronous, there
is a race condition that happens when an SSL handshake fails; you
may get a "connection closed" exception from Jetty before you get
the "SSLHandshakeException" that would eventually be thrown by
the HTTP client library.

When writing tests that involve intentional SSL failures, the
tests need to be tolerant of the different types of exceptions that
can occur.  This commit changes some tests that were not working
that way to be a bit more tolerant of the different possible
exception types, while still trying to be as specific as possible
and validate that the exception actually represents the
expected SSL failure.