Remove path options from hostconfig
The necessary AIO defaults have been added to beaker, but can only be used
with `type => aio` set. This commit removes them from the host config, but
this means we can only run them against master with `type => aio`.

The puppetpath is not yet correct in a released version of beaker (but has
been merged in, see https://github.com/puppetlabs/beaker/issues/660), so for
now add it to the options; we should remove it later once a new version of
beaker has been released.