(maint) Update packaging & acceptance tests to create puppet user
Previously, the puppet package created the puppet user and group. However, the
puppet agent does not require this user, since it runs as root. The puppet
user is only required for the master. The puppet-agent AIO packages will not
be creating the puppet user or group, so the puppetserver packages must now do
this and must also set proper permissions on certain directories. This commit
adds the necessary steps to the postinstall.

It also removes the `puppet_apply_as_puppet_user` method from the acceptance
helpers, which 1) is no longer required, and 2) since the puppet user is not
created before installing the puppetserver package, we couldn't complete this
step.