(MAINT) Add JARS_NO_REQUIRE=true to gem subcommand (for the Nteenth time)
When the `gem` subcommand fixes were reverted in b2f83a77, the behavior
of excluding the Bouncy Castle jars from the JRuby library loading
process was lost in the shuffle.

This is a problem because acceptance tests are failing, particularly the
`puppetserver gem cleanup` command which errors out as follows:

    [root@u4e4c61i909s8tj ~]# puppetserver gem cleanup
    ERROR:  Loading command: cleanup (RuntimeError)
            you might need to reinstall the gem which depends on the missing jar
    no such file to load -- org/bouncycastle/bcpkix-jdk15on/1.47/bcpkix-jdk15on-1.47 (LoadError)
    ERROR:  While executing gem ... (NoMethodError)
        undefined method `invoke_with_build_args' for nil:NilClass

This patch addresses the problem by re-introducing the JARS_NO_REQUIRE
behavior into master as part of the merge-up from stable.