(SERVER-338) Ensure netstat is available for service management
Without this patch the puppet server service fails to start at the
service management framework layer.  The failure is caused by the
service unit attempting to use netstat to check for listening ports and
netstat not being available.

This patch addresses the problem by expressing a dependency on
net-tools, which contains netstat so that netstat is always available
after package installation and before service management.