(SERVER-206) Cleanup based on PR comments
Remove the use of the class variable in the HttpClient ruby class
and use class methods instead. DRY up code terminating scripting
containers in the HTTP Client tests. Remove the puppet loading
in the JRuby Testutils. Make the terminate method for the master
a method on the JRubyPuppet interface so it can be called on the
JRubyPuppet object.