(SERVER-338) Use lein-ezbake 0.1.2 to resolve netstat issue
Without this patch the puppetserver service script, managed by ezbake,
requires netstat but does not express the dependency.  This is a problem
because it causes the service to fail on startup.  This patch addresses
the problem by updating lein-ezbake from 0.1.0, which does not express a
dependency on net-tools to 0.1.2, which does.

In addition, this patch removes the project specific dependency on Java
as per the discussion with Matthaus that resulted in the following
principle:

There are duplicates of the same dependency expressed statically in the
[ezbake package specifications][requires].  The duplication is confusing
and more difficult to coordinate and maintain over time.

The principle for dependencies is that those required to satisfy the
behavior of ezbake itself, or the system level service management
scripts, should be expressed statically in ezbake since those
dependencies will apply to all projects using ezbake and the resulting
scripts.  Application level dependencies should be specified in the
project specific ezbake.conf file such that individual applications are
not interdependent on common dependencies.

[requires]: https://github.com/puppetlabs/ezbake/blob/418e4815edf159c753e148395b3e725f8d202c17/resources/puppetlabs/lein-ezbake/template/foss/ext/redhat/ezbake.spec.erb#L5