(SERVER-360) Fix common_package_name_in_different_providers test
Without this patch the ruby puppet
`common_package_name_in_different_providers.rb` acceptance test is
failing because the createrepo command is not available on the system
under test.  This is causing the puppet-server acceptance tests to fail.

This patch resolves the problem by satisfying the dependency in the
pre-suite.  This course of action is being taken over satisfying the
dependency directly in the common_package_name_in_different_providers.rb
test because the state of the puppet code-base is volatile with the AIO
work.

This change will move the puppet-server tests closer to being green
while also leaving a clear indication of where the dependency
installation should be moved once the puppet code base is less
volatile.