(PE-7726) Stop client service at end of integration test
Previously, the client service started in this test was not stopped at the end
of the test. This commit ammends the setup of the client service to use
`with-app-with-config`, which shuts down automatically after everything within
it is done.

Also, the client service does not depend on the webserver service, so it
doesn't need the webserver service set up with it.