(PE-7726) Shutdown services if error is thrown during integration test
This commit makes the network partition integration test more robust by
wrapping the stopping of the storage service in a try/finally block, so that
even if an exception is thrown during the tests, it will be stopped and
subsequent tests won't fail with "Address already in use" errors.

Also, wrap the storage service app creation with the `check-for-errors!`
function, so that an exception will be thrown and the entire test will fail if
the storage service does not start properly.