(PE-7726) Add comments about race condition in integration test
Currently, we are using `Thread/sleep` with a time greater that the client's
polling interval in the network partition integration test. However, just
hardcoding a timeout is not very robust and introduces potential race
conditions. This commit adjusts the polling interval and the time used in the
`Thread/sleep`s to try to decrease that likelihood, and adds TODO comments
about this for us to address in the future.