(PE-4800) Update file-sync-client config schemas
Update the file-sync-client config schemas so that, instead
of a vector, the list of repos is now a map where each key is the
name of a repo, and each value is either a string containing
the target directory for that repo or a map. If a map, it contains
one key, :target-dir, which contains the target directory of the
repo.