(SERVER-386) Depend on puppet-agent AIO dev version
Without this patch the puppet-agent dependency is out of date.  This is
a problem because we need to track puppet-agent as closely as possible
to minimize the likelihood of issues stemming from divergence from the
underlying Puppet behavior.

This updated version should have no behavior changes relative to the
behavior of the puppet-agent version in 33f5cf9.