(maint) Remove Gemfile.lock
Without this patch the Gemfile.lock is committed to the repository,
which is a problem because the Gemfile has relaxed the specification of
dependency versions, particularly for beaker and the ability to install
from a Git HEAD.  This is a problem because the Gemfile.lock is no
longer a single source of truth and will be frequently in conflict.

puppet-server follows the lead of puppet with regard to branching,
merging and Gemfile dependency management.  Puppet deleted this file
from their repository back in 2012 in commit
c6b8e7fce004ea393f7821b6e576568cc77f035b, so this patch follows their
lead.