(SERVER-369) Work around beaker problems with AIO paths
Without this patch acceptance tests are failing with false-positives
because the new AIO bin directory specified in the [Puppet Installation
Layout][layout] is not in the PATH for the system under test.  The
symptom of this problem is `puppetserver` commands erroring out with an
exit status of 127 (command not found).  This patch addresses the
problem by modifying the PATH in the pre-suite.  Ideally this will be
fixed in QE-1891, and the pre-suite hack can be removed once that ticket
is closed.

[layout]: https://github.com/puppetlabs/puppet-specifications/blob/af825090dba94c56d332b0e61d73395eb6eb0535/file_paths.md#puppet-server