(PE-8135) Use bare repo for file sync client service
Previously, the file sync client service kept a git repo including the full
working tree. This repo does not need to be a full working copy, since no one
will be working out of it. This commit updates the file sync client to keep a
bare repo instead. Since it is now keeping a bare repo, it does a `git fetch`
rather than a `git pull` to get updates from the storage service.

This commit also adds helper functions for dealing with bare repos.
Specifically, it renames `get-repository` to
`get-repository-from-working-tree` and adds a function
`get-repository-from-git-dir` for use with bare directories that don't have
working trees. In addition, it changes the `head-rev-id` function to require a
repository instance, and adds a variant that takes a directory to be used as
the git dir and another that takes a directory to be used as the working tree.