(maint) refactor JGit client SSL configuration
Introduce a 'configure-jgit-client-ssl!' function in the file sync
client core namespace, with additional documentation on what this code
is doing and why it has to be sucky.  Also, re-wrote a test case for
this behavior to just test this function directly, instead of booting up
the entire service, resolving a problem with errors being logged during
the test because the service is started and then immediately shut down.