(PE-8135) Remove `pull` function from jgit-client
We have changed the file sync client to keep a bare repo and to update that
bare repo from the storage service using a `git fetch` rather than a `git
pull`. Consequently, the `pull` method is no longer needed, and, as far as we
currently know we will not have any need for it in the future, since we plan
to update the live code directory using the equivalent of a `git reset
--hard`. This commit removes the now-unused `pull` function from the
jgit-client namespace.