(PE-4783) basic job scheduling service
This commits adds a basic job scheduling service.  This will be used by
the file sync client to schedule its periodic synchronization of the
local git repository.  The implementation of the scheduler uses the
'at-at' library.

If an error is thrown from a scheduled job run by at-at, it is silently
swallowed.  The scheduler service implementation ensures that all errors
are logged by wrapping job functions with a try/catch block which logs
and re-throws any errors.