(SERVER-415) Update PUPPET_BUILD_VERSION
Without this patch the master branch of puppet-server runs acceptance
tests using an out of date version of the puppet-agent dependency.  This
is a problem because this out of date version decreases confidence that
the master branch of puppet-server is fully working with recent versions
of puppet-agent.

This patch addresses the problem by updating to build [dc53fa3][1], the
most recent build today, the day PUP-3997 and PUP-4056 have been merged
into puppet's master branch.

[1]: http://builds.puppetlabs.lan/puppet-agent/dc53fa36a1813454b9b55243a3f291dae122d614/