(PE-8163, PE-8210) introduce agent and fix tests
This commit ports the file sync client service implementation to use an
agent which manages the state of the sync process and ensures that only
one thread is ever running the sync process at a given point in time.
The agent has a simple state which represents the result of the last
iteration of the sync process, including any error that may have
occurred.

The agent is responsible for scheduling the subsequent iteration of the
sync process - thus, the 'after' function from the scheduler service is
now used instead of 'interspaced'.

As part of this work, the error handling was of the file sync client
service was re-worked; in particular, JGit errors are now
caught in a single place, and, in all cases, when an error state is
detected, a map is thrown via slingshot.  Also, when dealing with a response
from the 'latest commits' endpoint this commit simplifies the error
handling to be either a) the response is in a format we can handle,
or b) we just need to throw an error.

These changes allowed for lots of good improvements to the test code...

Race conditions are eliminated from the "network partition tolerance"
integration test.  This is accomplished by re-writing the test to
use watches which synchronously verify that the sync
process is doing what is expected, instead of terrible things like
'Thread/sleep' and scraping for log messages.

The client service test cases were also re-written to take advantage of
the agent's state, which eliminated race conditions and usages of 'with-redefs'.

Finally, this commit also consolidated the test helpers into a single namespace.