(SERVER-410) Always cache features in puppet server
Without this patch puppet server defers to puppet on the value of the
always_cache_features setting.  This is a problem because it does not
make sense for features to be reloaded frequently on the puppet server,
this behavior is only useful for puppet agent processes that need to
update the set of installed features mid-run.

This patch addresses the problem by explicitly setting
Puppet[:always_cache_features] to true in Puppet::Server::PuppetConfig.

This approach is employed to keep all of the static jruby-puppet related
settings in one place, internal to the JRuby contexts.  This approach is
employed in order to avoid expanding the surface area between Clojure
and JRuby.