(maint) Clear all AIO hacks and workarounds
Without this patch there are still hacks and work-arounds in place to
try and get the puppet-master master branch green while things were
still in flux in the puppet-agent dependency.

Things should no longer be in flux in puppet-agent now that PUP-3997 and
PUP-4056 have been merged into the master branch of puppet.  As a result
we need to remove all hacks and work-arounds to gain the most confidence
possible regarding the behavior of puppet-server and puppet-agent,
together, in peaceful harmony.