(SERVER-435) Pin beaker to 7d0cccf
Without this patch puppetserver is using an older version of beaker that
does not have all of the AIO path updates as of this point in time.
This patch updates beaker to version 7d0cccf which is the exact version
Puppet is using in their master branch as per PUP-4056.