(PE-4799) Add publish endpoint
Add endpoint to publish content from working directories to storage service.
When a POST is made to this endpoint, all unversioned files in the working
directory for each repo will be added, committed, and pushed up to the storage
service's bare repo. A commit will be made and pushed whether or not there are
unversioned files to add.

The body of the request can optionally contain a `message` and `author`, which
are used as the commit message and commit author when making the commit.

The endpoint will return a 400 error for requests that cannot be parsed as
JSON or that don't match the schema. Otherwise, a 200 with a JSON body
consisting of map of repo names to either a SHA for the latest commit on that
repo if the add/commit/push succeeded, or an error if there was a failure.