(maint) Check service start right out of the package
Without this patch the acceptance tests in the master branch are passing
with all green, successful, positive results.  This is a problem because
those results are misleading, immediately out of the package
puppetserver fails to start due to permission errors.

The problem is masked by the use of `puppet` in beaker shortly after
installing the puppetserver package.  The use of `puppet resource`,
`puppet apply` or `puppet agent` as root has the side-effect of changing
ownership and permissions of /var/run/puppetlabs, /var/log/puppetlabs,
and /etc/puppetlabs/puppet, all of which have the follow-on effect of
allowing the puppetserver service to start because the puppet user it
runs as is then able to create the directories it needs.

This patch addresses the problem by trying to start the puppetserver
service immediately after installing the package, just a like an user
would.  The test fails, like it should.