(SERVER-414) Manage rundir during service startup
Without this patch the puppetserver puppet rundir does not exist, which
is a problem because the parent directory of rundir is owned by root and
puppetserver runs as puppet which is unable to create the rundir.  This
prevents the service from starting up out of the package.

This patch addresses the problem by creating the rundir directory in a
manner owned by the same account puppetserver runs as.  This eliminates
the need for puppetserver to create the directory and allows
puppetserver to create files in the rundir directory.

NOTE: This does not cover the case of `puppetserver foreground` which
will still fail if executed before the service is started using the
service management framework.