(SERVER-448) Reduce the default value for # JRubies
There was a glaring flaw in the way that we had previously been
calculating a default value for the number of JRuby instances.
The formula was 'num-cpus + 2', which meant that on a really large
server with e.g. 32 cores, you'd get 34 JRuby instances.  That
many JRuby instances will require a very large heap size, but we
default to a 2G heap size, so users were hitting OutOfMemory errors.

This commit changes the formula.  We now use 'num-cpus - 1', with
a minimum value of 1 and a maximum value of 4.  We also log a
warning message if the user hasn't provided an explicit value for
this setting, and suggest that they do so.