(SERVER-415) Update PUPPET_BUILD_VERSION for RE-4141
Without this patch we are not using a puppet-agent version in CI that
contains the package provides improvements in RE-4141.  This is a
problem because we need to run puppetserver through CI in combination
with a puppet-agent that includes the RE-4141 improvements.

This patch addresses the problem by updating to puppet-agent build
[890a4a2](http://builds.puppetlabs.lan/puppet-agent/890a4a29148a3ae2997cb1cf2cefd329e6badf3d/)
which includes these changes.

This commit also updates the submodule to the same commit of puppet
included in the puppet-agent builds, which is the 4.0.0-rc1 tag in puppet.

See: [puppet.json](https://github.com/puppetlabs/puppet-agent/blob/890a4a29148a3ae2997cb1cf2cefd329e6badf3d/configs/components/puppet.json)