(PE-4799) Update storage service config to include working directory
Previously, repos were represented in the storage service config as an array
of maps, with each map consisting of a `sub-path` key representing it's name
and an `http-push-enabled-key`. This commit updates this config to now include
the working directory for each repo.

Thus, the `repos` config is now a map with keys being the repo names. The
values are each a map that must include the `working-dir` for the repo and may
optionally include the `http-push-enabled` key.