(SERVER-450) Add proper semver parsing to unit tests
Without this patch the unit tests are not properly parsing semantic
version strings according to the specification.  This is a problem for
RC versions of dependencies with versions such as "4.0.0-rc1"  This
patch addresses the problem by using a third party library written by
our very own Deepak designed to parse and manipulate semantic version
data.