(PE-4799) Fix handling of schema errors in middleware
Previously, all schema errors in the storage service were handled in the
`wrap-schema-errors` middleware by returning a 400 ("Bad Request"). However,
it  is possible that a schema might be caused by something going wrong on the
server, not by an invalid request body. This commit separates out the
validation of the request body and throws a `user-data-invalid` error if it is
not valid. This error is caught by the `wrap-schema-errors` middleware, and a
400 is returned for it. For all other schema errors, a 500 is returned.