(maint) Pin to beaker version 8909e35
Without this patch the beaker version, pinned to 2.7.1, is a problem
because it does not contain the behavior necessary to use the new puppet
collection repository layout scheme.

This patch addresses the problem by pinning to beaker version
8909e35a97b8c30b11b77b784a47c8d175cd4ec8 which implements the behavior
described in QENG-2096 and tries to satisfy dependencies out of PC1
first.

This change should have no change in behavior until the PC1 repository
exists because repositories are tried using a precedence ordering.  If
there are no dependencies available in PC1 then the current behavior is
employed.