(maint) Update puppet-agent dependency to 1166ad6
Without this patch puppetserver is installing the puppet-agent
dependency from the devel sub-directory of the builds server.  This is a
problem because we need to verify puppetserver and puppet-agent are both
installable from the new Puppet Collection 1 sub-directory of "PC1"

This patch addresses the problem by updating the pin of puppet-agent to
1166ad687883061f8f7133634144759d2b9eb37b which has been published into
the PC1 location.