(SERVER-535) Fix several bugs in j-p-service-test
Many (most?) of the configuration maps that were being used
in the jruby-puppet-service-tests were invalid, but this wasn't
surfacing in the test runs because of the way the mock
JRuby instances were being created.  This commit improves
the validation of the input to the mock fixture, and fixes all
of the test configurations.