(SERVER-533) Update branching document for 2.1
Without this patch we don't have a documented plan for how to handle
changes destined for 2.1.x in our current world where 1.1 and 2.0 have
not yet been released.  This is a problem because we need to start
active development on 2.1 before 1.1 and 2.0 are released, but we only
have stable and master for 1.1 and 2.0 respectively.

This patch addresses the problem by establishing a plan of action that
introduces the 2.1.x branch and how we're going to handle it as a
deviation from our normal workflow with stable and master.