(maint) Pessimistically pin to beaker 2.9.0
Without this patch puppetserver is pinned to a specific GIT sha of
beaker.  This is a problem because using a Git checkout has the side
effect of transient false positive failures as described in QENG-2138.
This patch addresses the problem by updating the dependency expression
to so-called pessimistic dependency that pins beaker to released
versions installed via rubygems, not git, in the 2.9.x series.