(maint) Use absolute URL to file path spec
Without this patch the URL to the file path specification document
refers to the master tree.  This is a problem because the master tree
will change over time but the release notes have a specific context of
the 2.0.0 release.  This patch addresses the problem by using a specific
SHA to refer to the state of the document at the time of the 2.0.0
release, making the URL refer to the same document over time.