(maint) Fix information about package naming
Without this patch the note about the package name changing form
puppet-server to puppetserver is incorrect.  Our package names have
always been puppetserver, it is the Maven artifact that has changed from
puppet-server to puppetserver.  This patch addresses the problem by
removing the section from the What's Changed area of the document.