(maint) Delete unnecessary test helpers
Delete all of the helper functions in the test-utils namespace which
contain assertions.  None of the assertions in these helper functions
are actually needed by the test cases in which they were used, and they
were not testing things that were relevant to the test cases from which
they were used, instead just testing the behavior in the helpers
themselves.