(SERVER-325) Don't use full server stack for flush int test
This commit tweaks the big scary max-requests-flush integration
test so that it doesn't boot the full puppet server stack, and
instead it just uses tk's `with-app-with-config` to boot the
minimum set of required services.  This should allow us,
in a subsequent commit, to mock the jruby instances for this
one test, to prevent the really extreme memory usage that the
test currently requires.