(SERVER-571) Add puppet-agent libdir to ruby-load-path
Without this patch Hiera backend plugins distributed as modules are not
functioning as expected in Puppet Server 2.0.  The expected behavior is
that the plugin is usable in puppetserver after a successful agent run
that has synchronized plugins.

This patch addresses the problem by adding the
`/opt/puppetlabs/puppet/cache/lib` path to the ruby-load-path setting
which lines up with the [Puppet Installation Layout][layout].

[layout]: http://goo.gl/BWyXpo