(maint) Overhaul acceptance README with vmpooler information
Without this patch the `acceptance/README.md` file is a bit out of date
as it focuses on the setup of locally managed VM's for the purpose of
acceptance tests.  This is a problem because maintaining local VM's was
previously required but no longer is with the introduction of the
vmpooler.  The vmpooler is a quicker and easier way to get started with
acceptance tests because the VM's being used are pre-configured and
actively maintained as part of the production CI system.  This patch
addresses the problem by focusing the setup documentation on the
vmpooler approach.