(maint) Simplify error handling
Simplify the error handling in sync-on-agent.  The only place an
AgentError can be thrown is get-latest-commits-from-server, so the call
to that function was moved up inside sync-on-agent.

This commit also cleans up the schemas in client-core a bit, and removes
the ! from process-repos-for-updates since it now returns a meaningful
value.