(maint) Stop accepting pushes over HTTP
Remove the ability for the storage service to accept pushes over HTTP.
There is currently no need for this - all changes to the repositories
are expected to come from the working directory on the local filesystem.

This removes the code and configuration for managing and testing the
http.receivepack setting on the storage service's repositories.

Several test cases needed to be updated to be compatible with this
change, either because they were explicitly pushing over HTTP, or
because they were cloning over HTTP and then pushing to origin - in this
case, the tests were updated to simply clone and push through the
filesystem - the fact that they were cloning over HTTP was incidental
and not relevant to the behavior of the test.

Finally, this results in a slightly different error message from JGit
when the servlet receives a push; however, the new error message is
reasonable and the exception class is the same (TransportException).
Although the Exception is now getting thrown at a different spot in
JGit, it still seems reasonable.