(PE-4798) Remove unnecessary push from publish
Update the publish endpoint to simply add and commit - no
push is needed since there is only a Git repository managed by the
service.

This commit also removes error handling for RepositoryNotFoundExceptions
from publish-repos, as the known error cases which previously triggered
this type of error now triggers a GitAPIException instead, due to the
implementation changes in publish-repo.  This change is reflected in
publish-content-endpoint-response-test.