(SERVER-598) Change master routes back to regex
In PR review, it was pointed out that the changes to make
the node/catalog/report and similar routes more precise (i.e.,
by forcing a match specifically for a single node segment
in the URL as opposed to tolerating any URL structure that
followed after the `/node/`, etc., prefixes) would cause
a slight change in behavior in terms of what kind of error
the server would return on an invalid route.

This commit changes the routes back to capturing any
path following the prefixes (using a regex), and passing
matching requests down through to the Ruby layer, so that
we'll get the same error behavior we had before.