(SERVER-614) Initialize timeouts from ClientOptions
This commit corrects a problem with the initialization of HTTP client
options which had been introduced by SERVER-530, the merge-up of
changes from 1.0.8 into the 2.1.x branch.  The merge-up had caused
HTTP requests to try to call the `setConnectTimeoutMilliseconds` and
`setIdleTimeoutMilliseconds` methods on the `RequestOptions` class.
`RequestOptions` does not support those methods; they can only
be configured at the client and not a per-request level.  To adjust
for this, this PR moves the invocation of the timeout methods to the
client level via `ClientOptions`.