(SERVER-548) Remove rather than map Accept: pson
Without this patch the behavior of the code is to map `s` or `pson` to
`binary` which is a problem because `pson` doesn't really make sense to
map to `binary`.  This patch addresses the problem the removing `pson`
instead of mapping it, which has the same expected behavior when mapping
valid Puppet 3 API requests to Puppet 4 formats.