(SERVER-548) Ignore Accept: pson and consolidate s,raw => binary
Without this patch the processing of the Accept header is a bit unwieldy
separated out over a handful of functions.  This patch addresses the
problem by consolidating all Accept header processing into a single
function, `map-accept-header`  The function ignores a value of `pson`
even though this value is filtered out in some Puppet 4 API requests and
maps s and raw to binary.

Addresses [JB's comment][1]

[1]: https://github.com/puppetlabs/puppet-server/pull/529#discussion_r30069840