(SERVER-548) Add puppet 3 auth.conf integration test
Without this patch we have no integration test that covers the behavior
of puppetserver with a customized auth.conf file when handling puppet 3
requests.  This patch addresses the problem by adding an integration
test that exercises a customized node and catalog API request at the
integration layer.

The integration test verifies a request from the client with certified
identity `localhost` can access the node and catalog resource for
`public`.  The tests also verify the same client cannot access node or
catalog resources for `private`.