(SERVER-547) Remove upgrade 404 message
The old ring handler which displays a friendly 404 error message is no
longer needed since we now have a legacy routes compatibility layer.
This commit removes the need for a map as the value of the master
service's route, since there is now only the single master service mount
point required. However, compatibility is maintained with the way the
master routes were configured in Puppet server 2.0, as map.