(QENG-2385) Fix Puppet 3 Compatibility job for Deb platforms
Without this patch the Puppet 3 Compatibility job is failing on Debian
platforms.  The root cause of this failure is that the underlying
package manager does not accept a version of `3.7.5`.  Working with QE,
they recommend we solve this issue by appending `-1puppetlabs1` to the
version coming from the PUPPET_LEGACY_VERSION environment variable, set
by the Job parameters.  This patch implements the recommendation.

I've verified the beaker command the job executes now passes
successfully for Ubuntu 14.04 hosts.