(SERVER-548) Stop executing assertions from functions
Without this patch the approach to testing modified Puppet 3 request
headers is less than ideal because the assertions are wrapped in a
function and then executed at a distance from the test case.  This patch
addresses the problem by using a ring middleware to stash the munged
request in an Atom and then return that stashed, munged request to the
caller.  This allows the munged request to be tested directly rather
than at a distance.  Conceptually, this patch implements a tap into the
function pipeline sending the request to the master-handler.