(SERVER-655) Fix Bouncy Castle jar load failure
This commit has changes which allow Puppet Server to boot without
JRuby's jopenssl trying to load its own Bouncy Castle jar.

Prior to JRuby 1.7.20, we set the JARS_NO_REQUIRE environment variable
onto the ScriptingContainer in order to instruct JRuby's
jar-dependencies to not try to load the Bouncy Castle jar - which we
intentionally exclude from the Puppet Server uberjar so that
'jvm-ssl-utils' can pull in its own version.

In JRuby 1.7.20's jopenssl, jar-dependencies was not being required
implicitly and so the JARS_NO_REQUIRE variable was ignored.  This
led Puppet Server to fail to boot because the Bouncy Castle jar
could not be loaded.

This commit has Puppet Server do an explicit require of
jar-dependencies so that the JARS_NO_REQUIRE variable can be honored
properly again.  It also sets JARS_REQUIRE, as that appears to be
the environment variable which jar-dependencies will eventually
replace JARS_NO_REQUIRE with.