(PE-8993) Add basic submodule awareness to storage service
Add two new settings, `submodules-dir` and `submodules-working-dir` to the
repo configuration on the storage service. If these are present (either both
can be present or neither can be), then when the publish endpoint is
triggered, also "publish" any subdirectories in the submodules-working-dir.
This will create a bare repo for eachsubdirectory and add it as a submodule to
the parent repo (if it previously did not exist as a submodule on the parent
repo), or add and commit it and then do a git pull to update it within the
parent repo if it was already a submodule of the parent repo.

Due to https://bugs.eclipse.org/bugs/show_bug.cgi?id=467611 we cannot have the
sumbodule git repo already exist within the parent repo working tree, thus we
have to have a separate setting for `submodules-working-dir`.