(SERVER-324) Add support for flushing a single environment
This commit extends our existing functionality around flushing
the Puppet environment cache.  Prior to this commit, the only
granularity that we allowed the cache to be flushed at was effectively
global.  This commit introduces functions for flushing an
individual environment, wires those functions all the way up through
the JRubyService protocol, and extends the puppet-admin-api HTTP endpoint
to support passing in an environment name as a query parameter.