(SERVER-657) Bump lein-ezbake dependency to 0.3.6
This commit bumps the lein-ezbake dependency to version 0.3.6.  This is
needed in order to allow Puppet Server packaging to include an updated
version of the foreground command which allows the HOME environment
variable to be set for the puppet user when the foreground subcommand is
executed.  Without this change, attempts to execute the foreground
subcommand on Ubuntu via sudo would fail because an inappropriate value
for the HOME environment variable was being used.